requirements.txt
pydantic
python-dotenv
openai
transitions
langchain
